// types/Candidato.ts
export interface Candidato {
  id: number;

  nombre: string;

  telefono: string;

  empresaId: number;

  vacanteId: number;

  municipio: string;

  edad: string;

  sexo: string;

  transporte: string;

  horaCita: string;

  lugarPresentacion: string;

  estatus: string;

  observaciones: string;
}